Move auth-related objects into separate module.
