+++ tiled/client/__init__.py
@@ -1,5 +1,5 @@
+from .auth import logout, logout_all, sessions  # noqa: F401
 from .constructors import from_context  # noqa: F401
 from .constructors import from_config, from_profile, from_tree, from_uri  # noqa: F401
-from .context import logout, logout_all, sessions  # noqa: F401
 from .node import ASCENDING, DESCENDING  # noqa: F401
 from .utils import hide_logs, record_history, show_logs  # noqa: F401
